+++ src/inspect_ai/_view/www/src/navbar/Navbar.mjs
@@ -19,6 +19,8 @@ import { SecondaryBar } from "./SecondaryBar.mjs";
  * @param {import("../types/log").EvalPlan} [props.evalPlan] - The EvalSpec
  * @param {import("../types/log").EvalStats} [props.evalStats] - The EvalStats
  * @param {import("../api/Types.mjs").SampleSummary[]} [props.samples] - the samples
+ * @param {import("../samples/SamplesDescriptor.mjs").EvalDescriptor} [props.evalDescriptor] - The EvalDescriptor
+ * @param {(fragment: string) => void} props.addToFilterExpression - add to the current filter expression
  * @param {string} [props.status] - the status
  * @param {boolean} props.offcanvas - Are we in offcanvas mode?
  * @param {boolean} props.showToggle - Should we show the toggle?
@@ -32,6 +34,8 @@ export const Navbar = ({
   evalResults,
   evalStats,
   samples,
+  evalDescriptor,
+  addToFilterExpression,
   showToggle,
   offcanvas,
   status,
@@ -182,6 +186,8 @@ export const Navbar = ({
           evalResults=${evalResults}
           evalStats=${evalStats}
           samples=${samples}
+          evalDescriptor=${evalDescriptor}
+          addToFilterExpression=${addToFilterExpression}
           status=${status}
           style=${{ gridColumn: "1/-1" }}
         />